requirements.txt
PyYAML==3.12
influxdb==5.0.0
requests==2.18.4
